requirements.txt
scikit-learn
numpy
Pillow
joblib
tk
pandas